changing monster value attribute to double (need to cast money to int)

+++ cookie_defender/map.h
@@ -59,8 +59,8 @@ private :
     Tower *t;
 
     int indexTower=0;
-    int health=20;
-    int stackHealth;
+    double health=20;
+    double stackHealth;
     int towerNumber;
     int width;
     int height;
